unroll the message-block-to-ints loop
I think there was some bounds checking in read_u64_into, which should've
been getting elided but wasn't. This makes a tiny difference in
performance -- maybe half a percent -- but it means that our portable
implementation finally pulls ahead of coreutils b2sum.

While we're at it, unroll some similar loops in lib.rs. Those don't
matter as much as stuff that happens in the compression function, but
we might as well be consistent.