allow hashing multiple inputs again with `blake2`
`b2sum` prints the filepath by default even for one argument, which
makes it inconvenient to use for examples in the Bao docs. When I
changed the name to `blake2`, I took the liberty of dropping the
filepath and restricting it to one argument. But multiple arguments is
convenient, so now I'm putting them back, but I only print the filepath
in the multi-arg case.