divorce the &Params lifetime from HashManyJob
There's a tricky situation where you have key bytes but an empty input
slice. That means the key block needs to be compressed with the
finalization flag. Since the compress loops unconditionally compress at
least one block (with that flag), we want to give them the key bytes as
input in that case, but that creates a lifetime relationship between the
params and the job.

I've run into a situation in Bao where I want to construct HashManyJobs
in a helper function and append them to a list, but I don't know all the
params until I'm in a loop inside the helper. That makes it surprisingly
tricky to initialize the Params in a stack spot that belongs to the
caller, but in such a way that rustc believes I'm not intializing them
twice. Rather than use some elaborate Option::take workaround for this
case, it makes sense to fix this lifetime restriction for all callers,
by simply finalizing the key block during construction in this case, and
skipping the subsequent call to the compression loop.

An alternative would be to stash an array of key bytes in every
HashManyJob, but that seems like an unnecessary performance hazard,
since the vast majority of callers won't have keys. We're already using
iterators to construct Job objects, so it's very easy to insert a
filter() for finished jobs.